toolkit/mozapps/downloads/tests/chrome/test_privatebrowsing_title.xul

<?xml version="1.0"?>
<!--
/* ***** BEGIN LICENSE BLOCK *****
 * Version: MPL 1.1/GPL 2.0/LGPL 2.1
 *
 * The contents of this file are subject to the Mozilla Public License Version
 * 1.1 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 * http://www.mozilla.org/MPL/
 *
 * Software distributed under the License is distributed on an "AS IS" basis,
 * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
 * for the specific language governing rights and limitations under the
 * License.
 *
 * The Original Code is Download Manager UI Test Code.
 *
 * The Initial Developer of the Original Code is
 * Ehsan Akhgari <ehsan.akhgari@gmail.com>.
 * Portions created by the Initial Developer are Copyright (C) 2009
 * the Initial Developer. All Rights Reserved.
 *
 * Contributor(s):
 *
 * Alternatively, the contents of this file may be used under the terms of
 * either the GNU General Public License Version 2 or later (the "GPL"), or
 * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
 * in which case the provisions of the GPL or the LGPL are applicable instead
 * of those above. If you wish to allow use of your version of this file only
 * under the terms of either the GPL or the LGPL, and not to allow others to
 * use your version of this file under the terms of the MPL, indicate your
 * decision by deleting the provisions above and replace them with the notice
 * and other provisions required by the GPL or the LGPL. If you do not delete
 * the provisions above, a recipient may use your version of this file under
 * the terms of any one of the MPL, the GPL or the LGPL.
 *
 * ***** END LICENSE BLOCK ***** */

/**
 * Test bug 464800 to make sure that the download manager window title is
 * properly cleared when transitioning into the private browsing mode.
 */
-->

<window title="Private Browsing transition title change test"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript"
          src="utils.js"/>

  <script type="application/javascript">
  <![CDATA[

function test()
{
  // The private browsing service might not be available
  let pbs = null;
  try {
    pbs = Cc["@mozilla.org/privatebrowsing;1"].getService(Ci.nsIPrivateBrowsingService);
  } catch (e) {
  }
  if (!pbs) {
    todo(false, "skip test - no private browsing supported");
    return;
  }

  var dmui = getDMUI();
  if (!dmui) {
    todo(false, "skip test for toolkit download manager UI");
    return;
  }

  // We should prevent the private browsing service from changing the current
  // session on private mode transitions, in order to play nice with the
  // Mochitest test harness.
  let prefBranch = Cc["@mozilla.org/preferences-service;1"].
                   getService(Ci.nsIPrefBranch);
  prefBranch.setBoolPref("browser.privatebrowsing.keep_current_session", true);

  // Start with an empty downloads database
  Cc["@mozilla.org/download-manager;1"].
  getService(Ci.nsIDownloadManager).
  DBConnection.
  executeSimpleSQL("DELETE FROM moz_downloads");

  // Close the UI if necessary
  let wm = Cc["@mozilla.org/appshell/window-mediator;1"].
           getService(Ci.nsIWindowMediator);
  let win = wm.getMostRecentWindow("Download:Manager");
  if (win) win.close();

  let obs = Cc["@mozilla.org/observer-service;1"].
            getService(Ci.nsIObserverService);
  const DLMGR_UI_DONE = "download-manager-ui-done";

  let testObs = {
    observe: function(aSubject, aTopic, aData) {
      if (aTopic != DLMGR_UI_DONE)
        return;
      obs.removeObserver(testObs, DLMGR_UI_DONE);

      let win = aSubject.QueryInterface(Ci.nsIDOMWindow);
      win.focus();
      let defaultTitle = win.document.documentElement.getAttribute("statictitle");
      is(win.document.title, defaultTitle,
        "The downloads window title is correct outside of the private browsing mode");
      win.document.title = "some title";

      // Enter the private browsing mode
      pbs.privateBrowsingEnabled = true;

      is(win.document.title, defaultTitle,
        "The downloads window title is correct inside the private browsing mode");
      win.document.title = "another title";

      // Leave the private browsing mode
      pbs.privateBrowsingEnabled = false;

      is(win.document.title, defaultTitle,
        "The downloads window title is correct after leaving the private browsing mode");

      // We're done!
      win.close();
      prefBranch.clearUserPref("browser.privatebrowsing.keep_current_session");
      SimpleTest.finish();
    }
  };
  obs.addObserver(testObs, DLMGR_UI_DONE, false);

  // Show the Download Manager UI
  dmui.show();

  SimpleTest.waitForExplicitFinish();
}

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
</window>
